+++ fs/dev.h
@@ -34,13 +34,13 @@ static inline dev_t_ dev_fake_from_real(dev_t dev) {
 #define DEV_CHAR 1
 
 struct dev_rtc {
-    intptr_t (*open)(int major, int minor, struct fd *fd);
+    int (*open)(int major, int minor, struct fd *fd);
     int (*close)(int major, int minor, struct fd *fd);
     ssize_t (*read)(void *buf, size_t count);
 };
 
 struct dev_ops {
-    intptr_t (*open)(int major, int minor, struct fd *fd);
+    int (*open)(int major, int minor, struct fd *fd);
     struct fd_ops fd;
 };
